Access Control

+++ app/Http/Controllers/PostsController.php
@@ -9,6 +9,16 @@ use DB; //direct data base
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -80,6 +90,12 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post =  Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/post')->with('error', 'Unauthorized Page');
+        }
+
         return view('posts.edit')->with('post', $post);
     }
 
@@ -115,6 +131,12 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/post')->with('error', 'Unauthorized Page');
+        }
+
         $post->delete();
 
         return redirect('/post')->with('success', 'Post Removed');
